+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Export formatter configuration via `.formatter.exs`.
   * Support single line usage (i.e. `async with a <- 1, do: a`).
   * Re-throw uncaught values (i.e. `async with _ <- throw(:foo), do: :ok`).
+  * Re-raise unrescued errors (i.e. `async with _ <- raise("ops"), do: :ok`).
 
 ## v0.2.2